requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pymongo==4.10.1
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.3
streamlit==1.41.1
pandas==2.2.3
